refactor: reorganize toast notifications for improved structure
- Create a new file for toast notifications to enhance organization
- Update imports in related files to reference the new toast notifications module
- Remove redundant toast notification functions from the toast component
- Ensure all tests and stories are updated to reflect the new structure

Closes #605.
