+++ .ai/plan/refactor-fast-refresh-exports-1.md
@@ -2,7 +2,7 @@
 goal: Refactor Component Exports for Fast Refresh Compliance
 version: 1.0
 date_created: 2025-10-01
-last_updated: 2025-10-03
+last_updated: 2025-10-04
 owner: marcusrbrown
 status: 'In Progress'
 tags: ['refactor', 'code-quality', 'fast-refresh', 'developer-experience']
@@ -81,11 +81,11 @@ Improve development experience by separating component exports from utility expo
 
 | Task | Description | Completed | Date |
 |------|-------------|-----------|------|
-| TASK-017 | Create `components/ui/toast-notifications.ts` with toastNotifications and toast exports | | |
-| TASK-018 | Update `components/ui/toast.tsx` to import and re-export utilities | | |
-| TASK-019 | Update consuming files to import toast utilities from new path if needed | | |
-| TASK-020 | Run tests to verify Toast component functionality | | |
-| TASK-021 | Verify lint warnings eliminated for toast.tsx:238,443,444 | | |
+| TASK-017 | Create `components/ui/toast-notifications.tsx` with toastNotifications and toast exports | ✅ | 2025-10-04 |
+| TASK-018 | Update `components/ui/toast.tsx` to remove toastNotifications exports (no re-export to fix Fast Refresh) | ✅ | 2025-10-04 |
+| TASK-019 | Update consuming files to import toast utilities from new path (toast.test.tsx, toast.stories.tsx) | ✅ | 2025-10-04 |
+| TASK-020 | Run tests to verify Toast component functionality (935 tests passed) | ✅ | 2025-10-04 |
+| TASK-021 | Verify lint warnings eliminated for toast.tsx:238,443,444 (7 warnings remaining, 3 Fast Refresh warnings eliminated) | ✅ | 2025-10-04 |
 
 ### Phase 6: Verification & Cleanup
