+++ components/ui/toast.stories.tsx
@@ -1,6 +1,6 @@
 import type {Meta, StoryObj} from '@storybook/react'
 import {Button} from './button'
-import {toastNotifications} from './toast'
+import {toastNotifications} from './toast-notifications'
 
 // Create a simple wrapper component for Storybook
 const ToastDemo = ({variant, title, description}: {variant: string; title: string; description: string}) => {
